pyproject.toml
[tool.poetry]
name = "django_liis"
version = "0.1.0"
description = "Тестовое задание для Junior Backend Developer (Django)"
authors = ["Ilya Moiseev <ilya.moiseev95@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.4"
djoser = "^2.1.0"
requests = "2.26.0"
django-filter = "2.4.0"
djangorestframework = "3.12.4"
djangorestframework-simplejwt = "4.8.0"
gunicorn = "20.0.4"
PyJWT = "2.1.0"
pytz = "2020.1"
sqlparse = "0.3.1"
six = "1.15.0"
flake8 = "^4.0.1"
psycopg2-binary = "^2.9.3"


[tool.poetry.dev-dependencies]
black = "^22.1.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
